+++ projects/net.splitcells.website.server/CHANGELOG.md
@@ -16,6 +16,8 @@ The changelog format can be found [here](../../src/main/md/net/splitcells/networ
 * **2022-09-04**:
   1. Render `DEVELOPMENT.md` of each project.
   2. Render `BUILD.md` of each project.
+  3. Render `LICENSE.md` of each project.
+  4. Render `NOTICE.md` of each project.
 * **2022-08-27**: The global changelog now widens the definition of events searched in other changelogs.
   Previously every list starting with just a date, was considered an event description.
   Now random strings after the date are accepted, as long as these are seperated by a whitespace from the date.